+++ index.php
@@ -5,11 +5,11 @@ Description: Collection of code to help developers customize WordPress into full
 Plugin URI: http://github.com/hrsetyono/edje-wp
 Author: The Syne Studio
 Author URI: http://thesyne.com/
-Version: 0.3.3
+Version: 0.3.4
 */
 
-require_once "lib/all.php";
-require_once "vendor/all.php";
+require_once 'lib/all.php';
+require_once 'vendor/all.php';
 
 // Main portal for calling all methods
 class H {
@@ -55,8 +55,8 @@ class H {
   public static function add_submenu($parent_title, $args) {
     $new_args = array(
       $parent_title => array(
-        "position" => "on {$parent_title}",
-        "submenu" => $args
+        'position' => "on $parent_title",
+        'submenu' => $args
       )
     );
     H::add_menus($new_args);
@@ -65,8 +65,8 @@ class H {
   public static function add_menu_counter($parent_title, $count_function) {
     $new_args = array(
       $parent_title => array(
-        "position" => "on {$parent_title}",
-        "counter" => $count_function,
+        'position' => "on $parent_title",
+        'counter' => $count_function,
       )
     );
 
@@ -77,24 +77,24 @@ class H {
 // ---------------
 // Github updater
 // ---------------
-add_action("init", "h_updater");
+add_action('init', 'h_updater');
 function h_updater() {
-  require_once "vendor/updater.php";
+  require_once 'vendor/updater.php';
 
   if (is_admin() ) {
-    $plugin_repo = "hrsetyono/edje-wp";
+    $plugin_repo = 'hrsetyono/edje-wp';
     $config = array(
-      "slug" => plugin_basename(__FILE__),
-      "proper_folder_name" => "edje-wp",
-      "api_url" => "https://api.github.com/repos/{$plugin_repo}",
-      "raw_url" => "https://raw.github.com/{$plugin_repo}/master",
-      "github_url" => "https://github.com/{$plugin_repo}",
-      "zip_url" => "https://github.com/{$plugin_repo}/archive/master.zip",
-      "sslverify" => true,
-      "requires" => "4.4.0",
-      "tested" => "4.4.0",
-      "readme" => "README.md",
-      "access_token" => "", // for private repo, authorize under Appearance > Github Update
+      'slug' => plugin_basename(__FILE__),
+      'proper_folder_name' => 'edje-wp',
+      'api_url' => "https://api.github.com/repos/$plugin_repo",
+      'raw_url' => "https://raw.github.com/$plugin_repo/master",
+      'github_url' => "https://github.com/$plugin_repo",
+      'zip_url' => "https://github.com/$plugin_repo/archive/master.zip",
+      'sslverify' => true,
+      'requires' => '4.4.0',
+      'tested' => '4.4.0',
+      'readme' => 'README.md',
+      'access_token' => '', // for private repo, authorize under Appearance > Github Update
      );
      new WP_GitHub_Updater($config);
   }