+++ lib/h-menu.php
@@ -10,14 +10,14 @@ class H_Menu {
     Add new menu items
   */
   public function add() {
-    add_action("admin_menu", array($this, "_add") );
+    add_action('admin_menu', array($this, '_add') );
   }
 
   /*
     Remove the admin menu by specifying the display text (case sensitive)
   */
   public function remove() {
-    add_action("admin_menu", array($this, "_remove") );
+    add_action('admin_menu', array($this, '_remove') );
   }
 
   /////////
@@ -33,31 +33,31 @@ class H_Menu {
       $icon = null;
 
       // get position if specified
-      if(isset( $value["position"] )) {
-        $position = $this->get_position($menu, $value["position"]);
+      if(isset( $value['position'] )) {
+        $position = $this->get_position($menu, $value['position']);
       }
 
-      if(isset( $value["icon"] )) {
-        $icon = "dashicons-" . str_replace("dashicons-", "", $value["icon"]);
+      if(isset( $value['icon'] )) {
+        $icon = 'dashicons-' . str_replace('dashicons-', '', $value['icon']);
       }
 
       // add top level menu if slug is specified
-      if(isset( $value["slug"] )) {
-        add_menu_page($title, $title, "manage_options",
-          $value["slug"], null, $icon, $position
+      if(isset( $value['slug'] )) {
+        add_menu_page($title, $title, 'manage_options',
+          $value['slug'], null, $icon, $position
         );
       }
 
       // if has submenu
-      if(isset( $value["submenu"] )) {
+      if(isset( $value['submenu'] )) {
         $parent_slug = $menu[$position][2];
-        
-        $smenu = new H_Submenu($parent_slug, $value["submenu"]);
+
+        $smenu = new H_Submenu($parent_slug, $value['submenu']);
         $smenu->add();
       }
 
       // If has counter
-      if(isset( $value["counter"] )) {
+      if(isset( $value['counter'] )) {
         $menu[$position][0] .= $this->add_counter($value["counter"]);
       }
     endforeach;
@@ -71,7 +71,7 @@ class H_Menu {
 
     foreach($menu as $index => $value):
       if($value[0]) {
-        $i = explode(" <", $value[0]); // sometimes has <span> HTML, so take the first one
+        $i = explode(' <', $value[0]); // sometimes has <span> HTML, so take the first one
 
         if(in_array($i[0], $args) ) {
           unset($menu[$index]);
@@ -90,7 +90,7 @@ class H_Menu {
   */
   private function add_counter($count_function) {
     $count = $count_function();
-    return " <span class='update-plugins count-{$count}'><span class='plugin-count'>{$count}</span></span>";
+    return " <span class='update-plugins count-$count'><span class='plugin-count'>$count</span></span>";
   }
 
 
@@ -98,13 +98,13 @@ class H_Menu {
     Look for available position based on anchor
 
     @param array $menu - The global $menu variable
-    @param array $raw_position - The stringified position, ex: "below Posts".
+    @param array $raw_position - The stringified position, ex: 'below Posts'.
 
     @return int - The position of anchor
   */
   private function get_position($menu, $raw_position) {
     // parse the string
-    preg_match("/(\w+)\s(\D+)/", $raw_position, $anchor);
+    preg_match('/(\w+)\s(\D+)/', $raw_position, $anchor);
     $placement = $anchor[1];
     $title = $anchor[2];
 
@@ -121,13 +121,13 @@ class H_Menu {
 
     // Set for increment or decrement
     switch($placement) {
-      case "above":
+      case 'above':
         $increment = -1;
         break;
-      case "below":
+      case 'below':
         $increment = 1;
         break;
-      case "on":
+      case 'on':
         return $position; // no need for increment
     }