+++ lib/post-filter.php
@@ -17,7 +17,7 @@ class H_PostFilter {
   }
 
   public function init() {
-    add_action("restrict_manage_posts", array($this, "my_restrict_manage_posts") );
+    add_action('restrict_manage_posts', array($this, 'my_restrict_manage_posts') );
   }
 
   /*
@@ -33,15 +33,16 @@ class H_PostFilter {
       // create an array of taxonomy slugs you want to filter by - if you want to retrieve all taxonomies, could use get_taxonomies() to build the list
       $filters = $this->post_taxonomies[$typenow];
 
-      foreach ($filters as $tax_slug) {
+      foreach ($filters as $tax) {
         // retrieve the taxonomy object
-        $tax_obj = get_taxonomy($tax_slug);
+        $tax_obj = get_taxonomy($tax);
         $tax_name = $tax_obj->labels->menu_name;
 
         // output html for taxonomy dropdown filter
-        echo "<select name='".strtolower($tax_slug)."' id='".strtolower($tax_slug)."' class='postform'>";
+        $tax_slug = strtolower($tax);
+        echo "<select name='$tax_slug' id='$tax_slug' class='postform'>";
         echo "<option value=''>Show All $tax_name</option>";
-        $this->generate_taxonomy_options($tax_slug,0,0,(isset($_GET[strtolower($tax_slug)])? $_GET[strtolower($tax_slug)] : null));
+        $this->generate_taxonomy_options($tax, 0, 0,(isset($_GET[$tax_slug])? $_GET[$tax_slug] : null));
         echo "</select>";
       }
     }
@@ -50,22 +51,23 @@ class H_PostFilter {
   /*
     Generate_taxonomy_options generate dropdown
   */
-  public function generate_taxonomy_options($tax_slug, $parent = "", $level = 0, $selected = null) {
-    $args = array("show_empty" => 1);
+  public function generate_taxonomy_options($tax_slug, $parent = '', $level = 0, $selected = null) {
+    $args = array('show_empty' => 1);
 
     if(!is_null($parent) ) {
-      $args = array("parent" => $parent);
+      $args = array('parent' => $parent);
     }
     $terms = get_terms($tax_slug, $args);
-    $tab = "";
+    $tab = '';
 
     for($i = 0; $i < $level; $i++) {
-      $tab .= "- ";
+      $tab .= '- ';
     }
 
     foreach ($terms as $term) {
       // output each select option line, check against the last $_GET to show the current option selected
-      echo "<option value=". $term->slug, $selected == $term->slug ? " selected='selected'" : "", ">" . $tab . $term->name ." (" . $term->count .")</option>";
+      $selected_attr = ($selected == $term->slug) ? ' selected="selected"' : '';
+      echo "<option value={$term->slug} $selected_attr>$tab {$term->name} ({$term->count})</option>";
       $this->generate_taxonomy_options($tax_slug, $term->term_id, $level + 1, $selected);
     }
   }